Removed bad events
 Changes to be committed:
	new file:   fabric-content-registries-v1/src/main/java/net/fabricmc/fabric/api/event/registry/v1/FabricRegistryEntryAddedEvents.java
	deleted:    fabric-content-registries-v1/src/main/java/net/fabricmc/fabric/api/event/registry/v1/RegistryBlockAddedCallback.java
	deleted:    fabric-content-registries-v1/src/main/java/net/fabricmc/fabric/api/event/registry/v1/RegistryBlockEntityAddedCallback.java
	deleted:    fabric-content-registries-v1/src/main/java/net/fabricmc/fabric/api/event/registry/v1/RegistryEntityAddedCallback.java
	deleted:    fabric-content-registries-v1/src/main/java/net/fabricmc/fabric/api/event/registry/v1/RegistryItemAddedCallback.java
	modified:   fabric-content-registries-v1/src/main/java/net/fabricmc/fabric/impl/content/registries/ContentRegistryImpl.java
